+++ packages/terra-badge/src/Badge.jsx
@@ -6,6 +6,10 @@ import styles from './Badge.module.scss';
 
 const cx = classNames.bind(styles);
 
+const BadgeScheme = ['default', 'primary', 'secondary', 'info', 'warning', 'positive', 'negative'];
+
+const BadgeSize = ['tiny', 'small', 'medium', 'large', 'huge'];
+
 const propTypes = {
   /**
    * Child Nodes.
@@ -18,7 +22,7 @@ const propTypes = {
   /**
    * Sets the badge color scheme. One of `default`, `primary`, `secondary`, `positive`, `negative`, `warning`, `info`.
    */
-  intent: PropTypes.oneOf(['default', 'primary', 'secondary', 'info', 'warning', 'positive', 'negative']),
+  intent: PropTypes.oneOf(BadgeScheme),
   /**
    * Reverses the position of the icon and text.
    */
@@ -26,7 +30,7 @@ const propTypes = {
   /**
    * Sets the badge size. One of `tiny`, `small`, `medium`, `large`, `huge`.
    */
-  size: PropTypes.oneOf(['tiny', 'small', 'medium', 'large', 'huge']),
+  size: PropTypes.oneOf(BadgeSize),
   /**
    * Sets the badge text.
    */
@@ -66,3 +70,4 @@ Badge.propTypes = propTypes;
 Badge.defaultProps = defaultProps;
 
 export default Badge;
+export { BadgeScheme, BadgeSize };