+++ packages/terra-text/src/Text.jsx
@@ -6,6 +6,10 @@ import styles from './Text.module.scss';
 
 const cx = classNames.bind(styles);
 
+const TextFontSize = [10, 12, 14, 16, 18, 20, 24, 32, 100];
+
+const TextWeight = [200, 400, 700];
+
 const propTypes = {
   /**
    * Child nodes
@@ -26,11 +30,11 @@ const propTypes = {
   /**
    * Sets the text size. One of `10`, `12`, `14`, `16`, `18`, `20`, `24`, `32`, `100`.
    */
-  fontSize: PropTypes.oneOf([10, 12, 14, 16, 18, 20, 24, 32, 100]),
+  fontSize: PropTypes.oneOf(TextFontSize),
   /**
    * Sets the text size. One of `200`, `400`, `700`.
    */
-  weight: PropTypes.oneOf([200, 400, 700]),
+  weight: PropTypes.oneOf(TextWeight),
 };
 
 const defaultProps = {
@@ -67,3 +71,4 @@ Text.propTypes = propTypes;
 Text.defaultProps = defaultProps;
 
 export default Text;
+export { TextFontSize, TextWeight };