+++ packages/terra-image/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Export for image variant added.
 
 2.21.1 - (September 6, 2018)
 ------------------